Created WOC Functionality and Updated GUI to have title
Created WOC Functionality and Updated GUI to have title.
WOC now takes into account the experts and takes the majority vote on if it should be filled or not. Seems to slightly outperform the best.
GUI keeps the c++ script running in the background meaning multiple graphs can be up at once. This has been reflected for WOC.